Rename HandlerGroup to SimpleHandlerGroup
