+++ yggdrash-core/src/main/java/io/yggdrash/core/net/PeerHandlerGroup.java
@@ -16,180 +16,28 @@
 
 package io.yggdrash.core.net;
 
-import io.yggdrash.core.blockchain.BlockHusk;
 import io.yggdrash.core.blockchain.BranchEventListener;
 import io.yggdrash.core.blockchain.BranchId;
-import io.yggdrash.core.blockchain.TransactionHusk;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
 
-public class PeerHandlerGroup implements BranchEventListener {
+public interface PeerHandlerGroup extends BranchEventListener {
 
-    private static final Logger log = LoggerFactory.getLogger(PeerHandlerGroup.class);
+    void setPeerEventListener(PeerEventListener peerEventListener);
 
-    private final Map<PeerId, PeerHandler> handlerMap = new ConcurrentHashMap<>();
+    PeerHandlerFactory getPeerHandlerFactory();
 
-    private PeerHandlerFactory peerHandlerFactory;
+    void destroyAll();
 
-    private PeerEventListener peerEventListener;
+    void healthCheck(Peer owner);
 
-    public PeerHandlerGroup(PeerHandlerFactory peerHandlerFactory) {
-        this.peerHandlerFactory = peerHandlerFactory;
-    }
+    int handlerCount();
 
-    public void setPeerEventListener(PeerEventListener peerEventListener) {
-        this.peerEventListener = peerEventListener;
-    }
+    void addHandler(Peer owner, Peer requestPeer);
 
-    PeerHandlerFactory getPeerHandlerFactory() {
-        return peerHandlerFactory;
-    }
+    List<String> getActivePeerList();
 
-    public void destroy() {
-        handlerMap.values().forEach(PeerHandler::stop);
-    }
+    List<String> getActivePeerListOf();
 
-    public void healthCheck(Peer owner) {
-        if (handlerMap.isEmpty()) {
-            log.trace("Active peer is empty to health check peer");
-            return;
-        }
-
-        for (PeerHandler handler : new ArrayList<>(handlerMap.values())) {
-            try {
-                String pong = handler.ping("Ping", owner);
-                if ("Pong".equals(pong)) {
-                    continue;
-                }
-            } catch (Exception e) {
-                log.warn(e.getMessage());
-            }
-            log.warn("Health check fail. peer=" + handler.getPeer().getYnodeUri());
-            removeHandler(handler);
-        }
-    }
-
-    @Override
-    public void receivedTransaction(TransactionHusk tx) {
-        if (handlerMap.isEmpty()) {
-            log.trace("Active peer is empty to broadcast transaction");
-            return;
-        }
-        for (PeerHandler peerHandler : handlerMap.values()) {
-            try {
-                peerHandler.broadcastTransaction(tx);
-            } catch (Exception e) {
-                removeHandler(peerHandler);
-            }
-        }
-    }
-
-    @Override
-    public void chainedBlock(BlockHusk block) {
-        if (handlerMap.isEmpty()) {
-            log.trace("Active peer is empty to broadcast block");
-            return;
-        }
-        for (PeerHandler peerHandler : handlerMap.values()) {
-            try {
-                peerHandler.broadcastBlock(block);
-            } catch (Exception e) {
-                removeHandler(peerHandler);
-            }
-        }
-    }
-
-    private void removeHandler(PeerHandler peerHandler) {
-        peerHandler.stop();
-        handlerMap.remove(peerHandler.getPeer().getPeerId());
-        peerEventListener.peerDisconnected(peerHandler.getPeer());
-        log.debug("Removed handler size={}", handlerMap.size());
-    }
-
-    int handlerCount() {
-        return handlerMap.size();
-    }
-
-    void addHandler(Peer owner, Peer requestPeer) {
-        PeerHandler peerHandler =  peerHandlerFactory.create(requestPeer);
-
-        if (handlerMap.containsKey(requestPeer.getPeerId())) {
-            return;
-        }
-
-        try {
-            log.info("Connecting... peer {}:{}", requestPeer.getHost(), requestPeer.getPort());
-            String pong = peerHandler.ping("Ping", owner);
-            // TODO validation peer
-            if ("Pong".equals(pong)) {
-                handlerMap.put(requestPeer.getPeerId(), peerHandler);
-                log.info("Added size={}, handler={}", handlerMap.size(), requestPeer.toAddress());
-            } else {
-                // 접속 실패 시 목록 및 버킷에서 제거
-                peerEventListener.peerDisconnected(requestPeer);
-            }
-        } catch (Exception e) {
-            log.warn("Fail to add to the peer handler err=" + e.getMessage());
-        }
-    }
-
-    public List<String> getActivePeerList() {
-        return handlerMap.values().stream()
-                .map(handler -> handler.getPeer().toString())
-                .collect(Collectors.toList());
-    }
-
-    public List<String> getActivePeerListOf() {
-        return handlerMap
-                .values()
-                .stream()
-                .map(c -> String.format("%s:%d", c.getPeer().getHost(), c.getPeer().getPort()))
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Sync block list.
-     *
-     * @param offset the offset
-     * @return the block list
-     */
-    public List<BlockHusk> syncBlock(BranchId branchId, long offset) {
-        if (handlerMap.isEmpty()) {
-            log.trace("Active peer is empty to sync block");
-            return Collections.emptyList();
-        }
-        // TODO sync peer selection policy
-        PeerId key = (PeerId) handlerMap.keySet().toArray()[0];
-        PeerHandler peerHandler = handlerMap.get(key);
-        List<BlockHusk> blockList = peerHandler.syncBlock(branchId, offset);
-        log.debug("Synchronize block offset={} receivedSize={}, from={}", offset, blockList.size(),
-                peerHandler.getPeer());
-        return blockList;
-    }
-
-    /**
-     * Sync transaction list.
-     *
-     * @return the transaction list
-     */
-    public List<TransactionHusk> syncTransaction(BranchId branchId) {
-        if (handlerMap.isEmpty()) {
-            log.trace("Active peer is empty to sync transaction");
-            return Collections.emptyList();
-        }
-        // TODO sync peer selection policy
-        PeerId key = (PeerId) handlerMap.keySet().toArray()[0];
-        PeerHandler peerHandler = handlerMap.get(key);
-        List<TransactionHusk> txList = peerHandler.syncTransaction(branchId);
-        log.info("Synchronize transaction receivedSize={}, from={}", txList.size(),
-                peerHandler.getPeer());
-        return txList;
-    }
+    List<PeerHandler> getHandlerList(BranchId branchId);
 }
